Graphics/Sprite: Fix warning in Sprite.inl
Former-commit-id: a3410af1934a462a4fe197700379ed64bbd6c76b [formerly 318faba08dbf1a4216a1fa5626be22c3b4973e92] [formerly 3f938befd96d748b6261b20301b8cb4afde287a5 [formerly 2dcf1a7bd016c4c81d9ecf08056ea202ac0c47ef]]
Former-commit-id: 742fa64e7f5954bca97ac7dd7ae11697fe47eea9 [formerly bd0de568838add11360d8dbd27af5c7ae0ddb1bb]
Former-commit-id: 35ca6c90855d4b3956e0cea3f7c265a286b121b5

+++ include/Nazara/Graphics/Sprite.inl
@@ -89,7 +89,7 @@ namespace Nz
 	*/
 	inline const Color& Sprite::GetCornerColor(RectCorner corner) const
 	{
-		NazaraAssert(corner < m_cornerColor.size(), "Invalid corner");
+		NazaraAssert(static_cast<std::size_t>(corner) < m_cornerColor.size(), "Invalid corner");
 
 		return m_cornerColor[corner];
 	}
@@ -164,7 +164,7 @@ namespace Nz
 	*/
 	inline void Sprite::SetCornerColor(RectCorner corner, const Color& color)
 	{
-		NazaraAssert(corner < m_cornerColor.size(), "Invalid corner");
+		NazaraAssert(static_cast<std::size_t>(corner) < m_cornerColor.size(), "Invalid corner");
 
 		m_cornerColor[corner] = color;
